Move <gazebo> exclusive elements to new gazebo parsers submodule
