+++ pcg_gazebo/parsers/gazebo/material.py
@@ -1,0 +1,28 @@
+# Copyright (c) 2020 - The Procedural Generation for Gazebo authors
+# For information on the respective copyright owner see the NOTICE file
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from ..types import XMLString
+
+
+class Material(XMLString):
+    _NAME = 'material'
+    _TYPE = 'gazebo'
+
+    def __init__(self, default='none'):
+        XMLString.__init__(self, default)
+
+    def to_sdf(self):
+        from ..sdf import Material
+        sdf = Material.get_gazebo_material(self._value)
+        return sdf
