+++ pcg_gazebo/parsers/gazebo/mu2.py
@@ -12,23 +12,33 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 from ..types import XMLScalar
 
 
-class StopCFM(XMLScalar):
-    _NAME = 'stopCfm'
-    _TYPE = 'urdf'
+class Mu2(XMLScalar):
+    _NAME = 'mu2'
+    _TYPE = 'gazebo'
 
-    def __init__(self, default=0):
+    def __init__(self, default=1):
         XMLScalar.__init__(self, default)
 
     def _set_value(self, value):
-        XMLScalar._set_value(self, value, min_value=0)
+        assert self._is_scalar(value), \
+            'Input value for {} must be a scalar'.format(self._NAME)
+        assert value >= 0, \
+            'Input value for {} must be equal' \
+            ' or greater than zero'.format(self._NAME)
+        XMLScalar._set_value(self, value)
 
-    def to_sdf(self):
+    def to_sdf(self, engine='ode'):
         from ..sdf import create_sdf_element
 
-        print('WARNING: stopCfm only translates to SDF cfm for the ODE engine')
-        obj = create_sdf_element('cfm')
+        assert engine in ['ode', 'bullet'], 'Accepted engine inputs are ode' \
+            ' or bullet'
+        if engine == 'ode':
+            obj = create_sdf_element('mu2')
+        else:
+            obj = create_sdf_element('friction2')
         obj.value = self.value
         return obj
