+++ pcg_gazebo/parsers/gazebo/provide_feedback.py
@@ -12,25 +12,18 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from ..types import XMLBoolean
 
-from ..types import XMLInteger
 
+class ProvideFeedback(XMLBoolean):
+    _NAME = 'provideFeedback'
+    _TYPE = 'gazebo'
 
-class MaxContacts(XMLInteger):
-    _NAME = 'maxContacts'
-    _TYPE = 'urdf'
-
-    def __init__(self, default=20):
-        XMLInteger.__init__(self, default)
-
-    def _set_value(self, value):
-        assert isinstance(value, int)
-        assert value > 0
-        XMLInteger._set_value(self, value)
+    def __init__(self, default=False):
+        XMLBoolean.__init__(self, default)
 
     def to_sdf(self):
         from ..sdf import create_sdf_element
-
-        obj = create_sdf_element('max_contacts')
+        obj = create_sdf_element('provide_feedback')
         obj.value = self.value
         return obj
