+++ pcg_gazebo/parsers/gazebo/stop_cfm.py
@@ -12,33 +12,23 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
 from ..types import XMLScalar
 
 
-class Mu2(XMLScalar):
-    _NAME = 'mu2'
-    _TYPE = 'urdf'
+class StopCFM(XMLScalar):
+    _NAME = 'stopCfm'
+    _TYPE = 'gazebo'
 
-    def __init__(self, default=1):
+    def __init__(self, default=0):
         XMLScalar.__init__(self, default)
 
     def _set_value(self, value):
-        assert self._is_scalar(value), \
-            'Input value for {} must be a scalar'.format(self._NAME)
-        assert value >= 0, \
-            'Input value for {} must be equal' \
-            ' or greater than zero'.format(self._NAME)
-        XMLScalar._set_value(self, value)
+        XMLScalar._set_value(self, value, min_value=0)
 
-    def to_sdf(self, engine='ode'):
+    def to_sdf(self):
         from ..sdf import create_sdf_element
 
-        assert engine in ['ode', 'bullet'], 'Accepted engine inputs are ode' \
-            ' or bullet'
-        if engine == 'ode':
-            obj = create_sdf_element('mu2')
-        else:
-            obj = create_sdf_element('friction2')
+        print('WARNING: stopCfm only translates to SDF cfm for the ODE engine')
+        obj = create_sdf_element('cfm')
         obj.value = self.value
         return obj
